Revert "Implemented core functionality"
