+++ calculatorUICOOP/MainPage.xaml.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Xamarin.Forms;
+
+namespace calculatorUICOOP
+{
+    // Learn more about making custom code visible in the Xamarin.Forms previewer
+    // by visiting https://aka.ms/xamarinforms-previewer
+    [DesignTimeVisible(false)]
+    public partial class MainPage : ContentPage
+    {
+        public MainPage()
+        {
+            InitializeComponent();
+
+
+
+        }
+        private void Button_Clicked(object sender, EventArgs e)
+        {
+            //storing the pressed button object 
+            var button = (Button)sender;
+           
+           
+        }
+    }
+}
