--- calculatorUICOOP/Models/Operator.cs
@@ -1,13 +1,0 @@
-﻿namespace calculatorUICOOP.Models
-{
-    /// <summary>
-    /// Mathmatical operator for: +, -, x, /
-    /// </summary>
-    public enum Operator
-    {
-        Add,        // +
-        Subtract,   // -
-        Multiply,   // x
-        Divide      // /
-    }
-}
