--- calculatorUICOOP/Static/Math.cs
@@ -1,20 +1,0 @@
-﻿namespace BasicXamarinCalc.Static
-{
-    /// <summary>
-    /// Generic Math class that evaluates two variables
-    /// </summary>
-    public static class Math
-    {
-        public static decimal Add(decimal x, decimal y) =>
-            x + y;
-
-        public static decimal Subtract(decimal x, decimal y) =>
-            x - y;
-
-        public static decimal Multiply(decimal x, decimal y) =>
-            x * y;
-
-        public static decimal Divide(decimal x, decimal y) =>
-            x / y;
-    }
-}
