--- calculatorUICOOP/Views/MainPage.xaml
@@ -1,63 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:d="http://xamarin.com/schemas/2014/forms/design"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             mc:Ignorable="d"
-             x:Class="calculatorUICOOP.MainPage">
-        
-                          <!--StaticResource styles numButton and opButton is in App.xaml -->
-                          <!-- Test commit -->
-
-    <Grid BackgroundColor="gray">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="30.1"/>
-            <RowDefinition Height="180"/>
-            <RowDefinition Height="100"/>
-            <RowDefinition Height="100"/>
-            <RowDefinition Height="100"/>
-            <RowDefinition Height="100"/>
-        </Grid.RowDefinitions>
-
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="*"/>
-            <ColumnDefinition Width="*"/>
-            <ColumnDefinition Width="*"/>
-        </Grid.ColumnDefinitions>
-
-        <Label  
-            BackgroundColor="DarkMagenta" Margin="-6" Text="  Calculator" TextColor="White" Grid.Row="0"
-            Grid.ColumnSpan="4" HorizontalTextAlignment="Start" VerticalTextAlignment="Center"  FontSize="20">
-        </Label>
-        <Label x:Name="DisplayLabel"
-            BackgroundColor="LightYellow" Text="0" Grid.Row="1" Grid.ColumnSpan="4" 
-            HorizontalTextAlignment="End" VerticalTextAlignment="Start" FontSize="50">
-        </Label>
-
-        <Button Clicked="Number_Clicked" Style="{StaticResource numButton}" Text="7" Grid.Row="2" Grid.Column="0"></Button>
-        <Button Clicked="Number_Clicked" Style="{StaticResource numButton}" Text="8" Grid.Row="2" Grid.Column="1"></Button>
-        <Button Clicked="Number_Clicked" Style="{StaticResource numButton}" Text="9" Grid.Row="2" Grid.Column="2"></Button>
-
-        <Button Clicked="Number_Clicked" Style="{StaticResource numButton}" Text="4" Grid.Row="3" Grid.Column="0"></Button>
-        <Button Clicked="Number_Clicked" Style="{StaticResource numButton}" Text="5" Grid.Row="3" Grid.Column="1"></Button>
-        <Button Clicked="Number_Clicked" Style="{StaticResource numButton}" Text="6" Grid.Row="3" Grid.Column="2"></Button>
-
-        <Button Clicked="Number_Clicked" Style="{StaticResource numButton}" Text="1" Grid.Row="4" Grid.Column="0"></Button>
-        <Button Clicked="Number_Clicked" Style="{StaticResource numButton}" Text="2" Grid.Row="4" Grid.Column="1"></Button>
-        <Button Clicked="Number_Clicked" Style="{StaticResource numButton}" Text="3" Grid.Row="4" Grid.Column="2"></Button>
-
-        <Button Clicked="Decimal_Clicked" Style="{StaticResource numButton}" Text="." Grid.Row="5" Grid.Column="0"></Button>
-        <Button Clicked="Number_Clicked" Style="{StaticResource numButton}" Text="0" Grid.Row="5" Grid.Column="1"></Button>
-        <Button Clicked="Equals_Clicked" Style="{StaticResource numButton}" Text="=" Grid.Row="5" Grid.Column="2"></Button>
-
-        <StackLayout Grid.RowSpan="4" Grid.Column="3" Grid.Row="2" BackgroundColor="LightGray" Padding="10" Margin="5">
-            <Button Clicked="C_Clicked" FontSize="20" Text="C" CornerRadius="10" BackgroundColor="DimGray" TextColor="White"></Button>
-            <Button Clicked="Operator_Clicked" Style="{StaticResource opButton}" Text="/"></Button>
-            <Button Clicked="Operator_Clicked" Style="{StaticResource opButton}" Text="x"></Button>
-            <Button Clicked="Operator_Clicked" Style="{StaticResource opButton}" Text="+"></Button>
-            <Button Clicked="Operator_Clicked" Style="{StaticResource opButton}" Text="-"></Button>
-            <Button Clicked="Percent_Clicked" Style="{StaticResource opButton}" Text="%"></Button>
-            <Button Clicked="CE_Clicked" Style="{StaticResource opButton}" Text="CE"></Button>
-        </StackLayout>
-    </Grid>
-</ContentPage>
